+++ RELEASENOTES.md
@@ -21,6 +21,11 @@
 
 # Release Notes
 
+### 5.0.4-hogangnono (Feb 28, 2024)
+
+- restore ios platform
+- set two splash screen for Ads
+
 ### 5.0.3-hogangnono (Jan 04, 2023)
 
 - removed android from supported platforms